SOPORTE_TECNICO_V2/SOPORTE_TECNICO_DIF/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
mysql-connector-python==8.2.0
Flask-Mail==0.10.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48
PyMySQL==1.1.2
bcrypt==5.0.0
python-dotenv==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.1.7
blinker==1.7.0
colorama==0.4.6
greenlet==3.0.3
typing-extensions==4.10.0
